% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_fude.R
\name{bind_fude}
\alias{bind_fude}
\title{Bind multiple Fude Polygon data objects}
\usage{
bind_fude(...)
}
\arguments{
\item{...}{Two or more Fude Polygon data objects to combine. Named elements that appear
in multiple inputs are row-bound into a single \code{\link[sf:sf]{sf::sf()}} object.}
}
\value{
A named list of \code{\link[sf:sf]{sf::sf()}} objects.
}
\description{
\code{bind_fude()} combines multiple Fude Polygon data objects by binding elements
with the same names across inputs. It can also be used on objects that have
already been combined by \code{\link[=combine_fude]{combine_fude()}}.
}
\examples{
path <- system.file("extdata", "castle.zip", package = "fude")
d1 <- read_fude(path, quiet = TRUE)
d2 <- read_fude(path, quiet = TRUE)
bind_fude(d1, d2)

}
\seealso{
\code{\link[=combine_fude]{combine_fude()}}
}
