% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_id_table}
\alias{city_id_table}
\title{City ID/name correspondence table}
\format{
A data frame with 1,970 rows and 9 variables:
\describe{
\item{key}{Unique city codes}
\item{pref_name}{Prefecture names in Kanji}
\item{pref_kana}{Prefecture names in Hiragana}
\item{pref_romaji}{Prefecture names in Romaji}
\item{city_name}{City names in Kanji}
\item{city_kana}{City names in Hiragana}
\item{city_romaji}{City names in Romaji}
\item{local_government_cd}{Local government codes}
\item{id}{ID}
}
}
\usage{
city_id_table
}
\description{
A dataset containing codes/names of cities in Japan.
}
\keyword{datasets}
