% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_year_id}
\alias{city_year_id}
\title{Community year/city ID correspondence table}
\format{
A data frame with 5,753 rows and 2 variables:
\describe{
\item{rcom_year}{Agricultural community reference year}
\item{id}{Reference ID for city_id_table}
}
}
\usage{
city_year_id
}
\description{
A dataset containing year/ID of communities in Japan.
}
\keyword{datasets}
