% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community_code_table}
\alias{community_code_table}
\title{Community code/name correspondence table}
\format{
A data frame with 149,511 rows and 13 variables:
\describe{
\item{KEY}{Unique community codes}
\item{PREF_NAME}{Prefecture names in Kanji}
\item{PREF_KANA}{Prefecture names in Hiragana}
\item{PREF_ROMAJI}{Prefecture names in Romaji}
\item{CITY_NAME}{City names in Kanji}
\item{CITY_KANA}{City names in Hiragana}
\item{CITY_ROMAJI}{City names in Romaji}
\item{KCITY_NAME}{Old village names in Kanji}
\item{RCOM_NAME}{Community names in Kanji}
\item{RCOM_KANA}{Community names in Hiragana}
\item{RCOM_ROMAJI}{Community names in Romaji}
\item{local_government_cd}{Local government codes}
\item{census_year}{The year of the census from which the data is derived}
}
}
\usage{
community_code_table
}
\description{
A dataset containing codes/names of communities in Japan.
}
\keyword{datasets}
