% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_fude.R
\name{filter_fude}
\alias{filter_fude}
\title{Get agricultural community boundary data}
\usage{
filter_fude(data, boundary, city, community, year = NULL)
}
\arguments{
\item{data}{List of \code{\link[sf:sf]{sf::sf()}} objects.}

\item{boundary}{List of one or more agricultural community boundary data provided by
the MAFF.}

\item{city}{A local government name in Japanese to be extracted. In the case of
overlapping local government names, the prefecture name must be included
(e.g., Fuchu-shi).}

\item{community}{String by regular expression. One or more agricultural community name in
Japanese to be extracted.}

\item{year}{Year in the column name of the \code{data}, if there is more than one
applicable local government code.}
}
\value{
A list of \code{\link[sf:sf]{sf::sf()}} objects.
}
\description{
\code{filter_fude()} uses the agricultural community boundary data to reduce the
Fude Polygon data to the community units.
}
\examples{
path <- system.file("extdata", "castle.zip", package = "fude")
d <- read_fude(path)
b <- get_boundary(d)
db <- filter_fude(d, b, "\u677e\u5c71\u5e02", "\u57ce\u6771", year = 2022)
}
