% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kcity_code_table}
\alias{kcity_code_table}
\title{Kcity code/name correspondence table}
\format{
A data frame with 12,110 rows and 10 variables:
\describe{
\item{KEY}{Unique kcity codes}
\item{PREF_NAME}{Prefecture names in Kanji}
\item{PREF_KANA}{Prefecture names in Hiragana}
\item{PREF_ROMAJI}{Prefecture names in Romaji}
\item{CITY_NAME}{City names in Kanji}
\item{CITY_KANA}{City names in Hiragana}
\item{CITY_ROMAJI}{City names in Romaji}
\item{KCITY_NAME}{Former village names in Kanji}
\item{local_government_cd}{Local government codes}
\item{census_year}{Year of the census from which the data is derived}
}
}
\usage{
kcity_code_table
}
\description{
A dataset containing codes/names of kcities in Japan.
}
\keyword{datasets}
