% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kcity_year_id}
\alias{kcity_year_id}
\title{Community year/Kcity ID correspondence table}
\format{
A data frame with 36,375 rows and 2 variables:
\describe{
\item{rcom_year}{Agricultural community reference year}
\item{id}{Reference ID for kcity_id_table}
}
}
\usage{
kcity_year_id
}
\description{
A dataset containing year/ID of communities in Japan.
}
\keyword{datasets}
