% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lg_code_table}
\alias{lg_code_table}
\title{Local government code/name correspondence table}
\format{
A data frame with 1,992 rows and 6 variables:
\describe{
\item{lg_code}{Local government codes}
\item{pref_kanji}{Prefecture names written in kanji}
\item{city_kanji}{Local government names written in kanji}
\item{pref_kana}{Prefecture names written in katakana}
\item{city_kana}{Local government names written in katakana}
\item{romaji}{Local government names written in romaji}
}
}
\usage{
lg_code_table
}
\description{
A dataset containing codes/names of local governments in Japan.
}
\keyword{datasets}
