% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fude.R
\name{ls_fude}
\alias{ls_fude}
\title{Inspect the Structure of Fude Polygon Data}
\usage{
ls_fude(data)
}
\arguments{
\item{data}{Fude Polygon data as returned by \code{\link[=read_fude]{read_fude()}}.}
}
\value{
A data frame.
}
\description{
\code{ls_fude()} lists the year and the local government names (or codes) in
order to understand what is included in the list returned by \code{\link[=read_fude]{read_fude()}}.
}
\seealso{
\code{\link[=read_fude]{read_fude()}}.
}
