% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pref_code_table}
\alias{pref_code_table}
\title{Prefecture code/name correspondence table}
\format{
A data frame with 47 rows and 4 variables:
\describe{
\item{pref}{Prefecture codes}
\item{pref_name}{Prefecture names in Kanji}
\item{pref_kana}{Prefecture names in Hiragana}
\item{pref_romaji}{Prefecture names in Romaji}
}
}
\usage{
pref_code_table
}
\description{
A dataset containing codes/names of prefectures in Japan.
}
\keyword{datasets}
