% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcom_id_table}
\alias{rcom_id_table}
\title{Community ID/name correspondence table}
\format{
A data frame with 154,069 rows and 13 variables:
\describe{
\item{key}{Unique community codes}
\item{pref_name}{Prefecture names in Kanji}
\item{pref_kana}{Prefecture names in Hiragana}
\item{pref_romaji}{Prefecture names in Romaji}
\item{city_name}{City names in Kanji}
\item{city_kana}{City names in Hiragana}
\item{city_romaji}{City names in Romaji}
\item{kcity_name}{Former village names in Kanji}
\item{rcom_name}{Community names in Kanji}
\item{rcom_kana}{Community names in Hiragana}
\item{rcom_romaji}{Community names in Romaji}
\item{local_government_cd}{Local government codes}
\item{id}{ID}
}
}
\usage{
rcom_id_table
}
\description{
A dataset containing codes/names of communities in Japan.
}
\keyword{datasets}
