% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcom_year_id}
\alias{rcom_year_id}
\title{Community year/ID correspondence table}
\format{
A data frame with 448,296 rows and 2 variables:
\describe{
\item{rcom_year}{Agricultural community reference year}
\item{id}{Reference ID for rcom_id_table}
}
}
\usage{
rcom_year_id
}
\description{
A dataset containing year/ID of communities in Japan.
}
\keyword{datasets}
