% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shuraku.R
\name{read_shuraku}
\alias{read_shuraku}
\title{Read shuraku data}
\usage{
read_shuraku(path, boundary, na = c("-", "…"), zero = TRUE)
}
\arguments{
\item{path}{Path to an \code{.xlsx} file.}

\item{boundary}{Agricultural community boundary data as returned by \code{\link[=get_boundary]{get_boundary()}}.}

\item{na}{Character vector of strings to interpret as missing values. Defaults to
\code{c("-", "\\u2026")}.}

\item{zero}{Logical. If \code{TRUE}, treat masked values (\code{"x"} and \code{"X"}) as zero.}
}
\value{
An \code{\link[sf:sf]{sf::sf()}} object.
}
\description{
\code{read_shuraku()} reads shuraku Excel files provided by MAFF.
}
