% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_code_table}
\alias{city_code_table}
\title{City code/name correspondence table}
\format{
A data frame with 1,905 rows and 9 variables:
\describe{
\item{KEY}{Unique kcity codes}
\item{PREF_NAME}{Prefecture names in Kanji}
\item{PREF_KANA}{Prefecture names in Hiragana}
\item{PREF_ROMAJI}{Prefecture names in Romaji}
\item{CITY_NAME}{City names in Kanji}
\item{CITY_KANA}{City names in Hiragana}
\item{CITY_ROMAJI}{City names in Romaji}
\item{local_government_cd}{Local government codes}
\item{census_year}{Year of the census from which the data is derived}
}
}
\usage{
city_code_table
}
\description{
A dataset containing codes/names of cities in Japan.
}
\keyword{datasets}
