% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_boundary.R
\name{extract_boundary}
\alias{extract_boundary}
\title{Extract specified agricultural community boundary data}
\usage{
extract_boundary(boundary, city = "", kcity = "", rcom = "", layer = FALSE)
}
\arguments{
\item{boundary}{Agricultural community boundary data returned by \code{\link[=get_boundary]{get_boundary()}}.}

\item{city}{A character vector of municipality names or local government codes used to
identify target municipalities. If \code{NULL}, all municipalities are kept.}

\item{kcity}{A character vector of regular expression patterns used to match former
municipality names in Japanese.}

\item{rcom}{A character vector of regular expression patterns used to match agricultural
community names in Japanese.}

\item{layer}{Logical. If \code{TRUE}, return a list containing extracted agricultural community
boundaries together with former municipality, municipality, and prefecture
boundary layers.}
}
\value{
If \code{layer = FALSE}, an \code{\link[sf:sf]{sf::sf()}} object. If \code{layer = TRUE}, a named list of
\code{\link[sf:sf]{sf::sf()}} objects.
}
\description{
\code{extract_boundary()} extracts subsets of agricultural community boundary data
returned by \code{\link[=get_boundary]{get_boundary()}} by municipality, former municipality, and/or
agricultural community.
}
\seealso{
\code{\link[=get_boundary]{get_boundary()}}
}
