% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fude.R
\name{extract_fude}
\alias{extract_fude}
\title{Extract specified Fude Polygon data}
\usage{
extract_fude(data, year = NULL, city = NULL, kcity = "", community = "")
}
\arguments{
\item{data}{A list of \code{\link[sf:sf]{sf::sf()}} objects.}

\item{year}{One or more years to extract.}

\item{city}{Local government names or codes to extract.}

\item{kcity}{A regular expression string. One or more former village names (in Japanese)
to extract.}

\item{community}{A regular expression string. One or more agricultural community names (in
Japanese) to extract.}
}
\value{
A list of \code{\link[sf:sf]{sf::sf()}} object(s).
}
\description{
\code{extract_fude()} extracts the specified data from the list returned by
\code{\link[=read_fude]{read_fude()}}.
}
\seealso{
\code{\link[=read_fude]{read_fude()}}.
}
