% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fude.R
\name{extract_fude}
\alias{extract_fude}
\title{Extract specified Fude Polygon data}
\usage{
extract_fude(data, year = NULL, city = NULL, kcity = "", rcom = "")
}
\arguments{
\item{data}{Fude Polygon data as returned by \code{\link[=read_fude]{read_fude()}}.}

\item{year}{A numeric vector of years to extract.}

\item{city}{A character vector of local government names or 6-digit local government
codes to extract.}

\item{kcity}{A regular expression. One or more former municipality names (in Japanese)
to extract.}

\item{rcom}{A regular expression. One or more agricultural community names (in
Japanese) to extract.}
}
\value{
An \code{\link[sf:sf]{sf::sf()}} object.
}
\description{
\code{extract_fude()} extracts specified subsets of Fude Polygon data returned by
\code{\link[=read_fude]{read_fude()}}.
}
\seealso{
\code{\link[=read_fude]{read_fude()}}.
}
