% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundary.R
\name{get_boundary}
\alias{get_boundary}
\title{Get the agricultural community boundary data}
\usage{
get_boundary(
  data,
  year = 2020,
  census_year = 2020,
  path = NULL,
  to_wgs84 = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data}{List of \code{\link[sf:sf]{sf::sf()}} objects or one or more strings representing prefecture
codes.}

\item{year}{The year when the agricultural community boundary data was created.}

\item{census_year}{The year of the Agricultural and Forestry Census.}

\item{path}{Path to the ZIP file containing the agricultural community boundary data;
use a local ZIP file instead of going looking for a ZIP file. Specify a
directory containing one or more ZIP files, not the ZIP file itself.}

\item{to_wgs84}{logical. Convert JGD2000 to WGS 84.}

\item{quiet}{logical. Suppress information about the data to be read.}
}
\value{
A list of \code{\link[sf:sf]{sf::sf()}} objects.
}
\description{
\code{get_boundary()} downloads and reads one or more agricultural community
boundary data provided by the MAFF.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
path <- system.file("extdata", "castle.zip", package = "fude")
d <- read_fude(path)
b <- get_boundary(d)
\dontshow{\}) # examplesIf}
}
