% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kcity_code_table}
\alias{kcity_code_table}
\title{Kcity code/name correspondence table}
\format{
A data frame with 12,110 rows and 10 variables:
\describe{
\item{key}{Unique kcity codes}
\item{pref_name}{Prefecture names in Kanji}
\item{pref_kana}{Prefecture names in Hiragana}
\item{pref_romaji}{Prefecture names in Romaji}
\item{city_name}{City names in Kanji}
\item{city_kana}{City names in Hiragana}
\item{city_romaji}{City names in Romaji}
\item{kcity_name}{Former city names in Kanji}
\item{local_government_cd}{Local government codes}
\item{census_year}{Year of the census from which the data is derived}
}
}
\usage{
kcity_code_table
}
\description{
A dataset containing codes/names of kcities in Japan.
}
\keyword{datasets}
