% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fude.R
\name{ls_fude}
\alias{ls_fude}
\title{Itemize the structure of Fude Polygon data}
\usage{
ls_fude(data)
}
\arguments{
\item{data}{List of \code{\link[sf:sf]{sf::sf()}} objects.}
}
\value{
A data.frame.
}
\description{
\code{ls_fude()} lists the year and the local government names (or codes) in
order to understand what is included in the list returned by \code{\link[=read_fude]{read_fude()}}.
}
\examples{
path <- system.file("extdata", "castle.zip", package = "fude")
d <- read_fude(path, stringsAsFactors = FALSE, quiet = TRUE)
ls_fude(d)

}
\seealso{
\code{\link[=read_fude]{read_fude()}}.
}
