% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fude.R
\name{shiny_fude}
\alias{shiny_fude}
\title{Generate Citation Text for Fude Polygon Data}
\usage{
shiny_fude(data)
}
\arguments{
\item{data}{A list or data.frame containing Fude Polygon data.}
}
\value{
A list with two elements: \code{ja} for Japanese citation text and \code{en}
for English citation text.
}
\description{
This function generates citation text in Japanese and English for Fude
Polygon Data.
}
