% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_fude.R
\name{shiny_fude}
\alias{shiny_fude}
\title{Prepare Leaflet Map for Fude Polygon Data}
\usage{
shiny_fude(data, community = FALSE)
}
\arguments{
\item{data}{A list or data frame containing Fude Polygon data.}

\item{community}{A logical value indicating whether to overlay community data on the map.}
}
\value{
A Leaflet map object with Fude Polygon data with an HTML table.
}
\description{
Prepares a Leaflet map for Fude Polygon data.
}
